backend/requirements.txt
Django
djangorestframework
gunicorn
psycopg2-binary
dj-database-url
django-cors-headers
djangorestframework-simplejwt
